Moved tab order logic to taborder.js in src/core/utils
